TestURIBuilder was getting excluded by Savant, this might be cleaner anyways
Pull the nested classes out, such that:
1) We don't have to craft a regex to deal with including RequestResult$TestURIBuilder.class, but still exclude the other tests
2) RequestResult is not so big
