+++ src/test/java/org/primeframework/mvc/test/RequestResult.java
@@ -28,7 +28,6 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
@@ -44,7 +43,6 @@ import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.node.JsonNodeType;
 import com.google.inject.Injector;
 import io.fusionauth.http.Cookie;
 import io.fusionauth.http.Cookie.SameSite;
@@ -75,7 +73,6 @@ import org.primeframework.mvc.message.l10n.MissingMessageException;
 import org.primeframework.mvc.message.scope.CookieFlashScope;
 import org.primeframework.mvc.security.Encryptor;
 import org.primeframework.mvc.util.CookieTools;
-import org.primeframework.mvc.util.QueryStringBuilder;
 import org.primeframework.mvc.util.QueryStringTools;
 import org.primeframework.mvc.util.ThrowingFunction;
 import org.primeframework.mvc.util.ThrowingRunnable;
@@ -1114,13 +1111,13 @@ public class RequestResult {
    * @param consumer The consumer to accept a URI builder to add parameters
    * @return This.
    */
-  public RequestResult assertRedirect(String uri, Consumer<TestURIBuilder> consumer) {
+  public RequestResult assertRedirect(String uri, Consumer<URIBuilder> consumer) {
     String redirect = response.getHeader(Headers.Location);
     if (redirect == null) {
       throw new AssertionError("\nActual redirect was null. Why do you want to assert on it? Status code was [" + response.getStatus() + "]");
     }
 
-    TestURIBuilder builder = TestURIBuilder.builder(uri);
+    URIBuilder builder = URIBuilder.builder(uri);
     consumer.accept(builder);
 
     String expectedUri = builder.build();
@@ -2059,486 +2056,4 @@ public class RequestResult {
     void accept(T t) throws Exception;
   }
 
-  public static class DOMHelper {
-    public String body;
-
-    public Document document;
-
-    public DOMHelper(String body, Document document) {
-      this.body = body;
-      this.document = document;
-    }
-
-    /**
-     * Perform any custom modifications to the HTML document
-     *
-     * @return this.
-     */
-    public DOMHelper custom(ThrowingRunnable runnable) throws Exception {
-      runnable.run();
-      return this;
-    }
-
-    /**
-     * Perform any custom modifications to the HTML document
-     *
-     * @param consumer the HTML document consumer
-     * @return this.
-     */
-    public DOMHelper custom(ThrowingConsumer<Document> consumer) throws Exception {
-      consumer.accept(document);
-      return this;
-    }
-
-    /**
-     * Remove an attribute from a DOM element.
-     *
-     * @param selector the DOM selector
-     * @param name     the name of the attribute to remove
-     * @return this.
-     */
-    public DOMHelper removeAttribute(String selector, String name) {
-      Element element = document.selectFirst(selector);
-      if (element == null) {
-        throw new AssertionError("Expected at least one element to match the selector " + selector + ". Found [0] elements instead. Unable to set element value.\n\nActual body:\n" + body);
-      }
-
-      element.removeAttr(name);
-
-      return this;
-    }
-
-    /**
-     * Set an attribute w/ value on a DOM element.
-     *
-     * @param selector the DOM selector
-     * @param name     the name of the attribute
-     * @param value    the value of the attribute
-     * @return this.
-     */
-    public DOMHelper setAttribute(String selector, String name, String value) {
-      Element element = document.selectFirst(selector);
-      if (element == null) {
-        throw new AssertionError("Expected at least one element to match the selector " + selector + ". Found [0] elements instead. Unable to set element value.\n\nActual body:\n" + body);
-      }
-
-      element.attr(name, value);
-
-      return this;
-    }
-
-    public DOMHelper setChecked(String selector, boolean value) {
-      Element element = document.selectFirst(selector);
-      if (element == null) {
-        throw new AssertionError("Expected at least one element to match the selector " + selector + ". Found [0] elements instead. Unable to set element value.\n\nActual body:\n" + body);
-      }
-
-      if (element.is("input[type=radio]") && value) {
-        Elements elements = document.select(element.tagName().toLowerCase() + "[type=radio][name=" + element.attr("name") + "]");
-        for (Element e : elements) {
-          e.attr("checked", false);
-        }
-      }
-
-      element.attr("checked", value);
-      return this;
-    }
-
-    public DOMHelper setValue(String selector, Object value) {
-      if (value != null) {
-        Element element = document.selectFirst(selector);
-        if (element == null) {
-          throw new AssertionError("Expected at least one element to match the selector " + selector + ". Found [0] elements instead. Unable to set element value.\n\nActual body:\n" + body);
-        }
-
-        // Handle a select element
-        if (element.is("select")) {
-          // Remove the selected attribute for each option, add it to the one that matches the requested value.
-          for (Element option : element.getElementsByTag("option")) {
-            if (option.attr("value").equals(value.toString())) {
-              option.attr("selected", "selected");
-            } else {
-              option.removeAttr("selected");
-            }
-          }
-        } else {
-          element.val(value.toString());
-        }
-      }
-
-      return this;
-    }
-  }
-
-  public static class HTMLAsserter {
-    public Document document;
-
-    public RequestResult requestResult;
-
-    public HTMLAsserter(RequestResult requestResult) {
-      this.requestResult = requestResult;
-      document = Jsoup.parse(requestResult.getBodyAsString());
-    }
-
-    /**
-     * Ensure a single element matches the provided selector.
-     *
-     * @param selector the DOM selector
-     * @param expected the number of expected matches for this selector in the DOM.
-     * @return this.
-     */
-    public HTMLAsserter assertElementCount(String selector, int expected) {
-      Elements elements = document.select(selector);
-      if (elements.size() != expected) {
-        throw new AssertionError("Expected [" + expected + "] elements to match the selector " + selector + " but found [" + elements.size() + "] elements instead." + ((elements.size() == 0) ? "" : "\n\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Ensure no elements match the provided selector.
-     *
-     * @param selector the DOM selector
-     * @return this.
-     */
-    public HTMLAsserter assertElementDoesNotExist(String selector) {
-      Elements elements = document.select(selector);
-      if (elements.size() > 0) {
-        throw new AssertionError("Expected 0 elements to match the selector " + selector + ". Found [" + (elements.size() + "] elements.\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Assert an element does not have an attribute. The value is not checked.
-     *
-     * @param selector  the DOM selector
-     * @param attribute the attribute you expect
-     * @return this.
-     */
-    public HTMLAsserter assertElementDoesNotHaveAttribute(String selector, String attribute) {
-      Element element = selectExpectOne(selector);
-
-      if (element.hasAttr(attribute)) {
-        throw new AssertionError("Expected the element not to have attribute [" + attribute + "]." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Ensure a single element matches the provided selector.
-     *
-     * @param selector the DOM selector
-     * @return this.
-     */
-    public HTMLAsserter assertElementExists(String selector) {
-      selectExpectOne(selector);
-      return this;
-    }
-
-    /**
-     * Assert an element has an attribute. The value is not checked.
-     *
-     * @param selector  the DOM selector
-     * @param attribute the attribute you expect
-     * @return this.
-     */
-    public HTMLAsserter assertElementHasAttribute(String selector, String attribute) {
-      Element element = selectExpectOne(selector);
-
-      if (!element.hasAttr(attribute)) {
-        throw new AssertionError("Expected the element to have attribute [" + attribute + "]." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Assert the element has an attribute with a specific value.
-     *
-     * @param selector  the DOM selector
-     * @param attribute the name of the attribute you expect
-     * @param value     the value of the attribute you expect
-     * @return this.
-     */
-    public HTMLAsserter assertElementHasAttributeValue(String selector, String attribute, String value) {
-      Element element = selectExpectOne(selector);
-
-      if (!element.hasAttr(attribute)) {
-        throw new AssertionError("Expected the element to have attribute [" + attribute + "]." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      String actual = element.attr(attribute);
-      if (!value.equals(actual)) {
-        throw new AssertionError("Attribute [" + attribute + "] value not equal to expected.\nExpected [" + value + "] but found [" + actual + "]" + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Ensure a single element matches the provided selector and has an expected inner HTML.
-     *
-     * @param selector          the DOM selector
-     * @param expectedInnerHTML the expected inner HTML
-     * @return this.
-     */
-    public HTMLAsserter assertElementInnerHTML(String selector, String expectedInnerHTML) {
-      Elements elements = document.select(selector);
-      if (elements.size() != 1) {
-        throw new AssertionError("Expected a single element to match the selector " + selector + ". Found [" + elements.size() + "] elements instead." + ((elements.size() == 0) ? "" : "\n\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      Element element = elements.get(0);
-      if (!expectedInnerHTML.equals(element.html())) {
-        throw new AssertionError("Expected a value of [" + expectedInnerHTML + "] to match the selector " + selector + ". Found [" + element.html() + "] instead." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Ensure a single element matches the provided selector and is "checked"
-     *
-     * @param selector the DOM selector
-     * @return this.
-     */
-    public HTMLAsserter assertElementIsChecked(String selector) {
-      Elements elements = document.select(selector);
-      if (elements.size() != 1) {
-        throw new AssertionError("Expected a single element to match the selector " + selector + ". Found [" + elements.size() + "] elements instead." + ((elements.size() == 0) ? "" : "\n\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      if (!elements.get(0).hasAttr("checked")) {
-        throw new AssertionError("Expected the element to be checked." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Ensure a single element matches the provided selector and is NOT "checked"
-     *
-     * @param selector the DOM selector
-     * @return this.
-     */
-    public HTMLAsserter assertElementIsNotChecked(String selector) {
-      Elements elements = document.select(selector);
-      if (elements.size() != 1) {
-        throw new AssertionError("Expected a single element to match the selector " + selector + ". Found [" + elements.size() + "] elements instead." + ((elements.size() == 0) ? "" : "\n\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      if (elements.get(0).hasAttr("checked")) {
-        throw new AssertionError("Expected the element NOT to be checked." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Ensure a single element matches the provided selector and equals the provided value.
-     *
-     * @param selector the DOM selector
-     * @param value    the expected value
-     * @return this.
-     */
-    public HTMLAsserter assertElementValue(String selector, Object value) {
-      Elements elements = document.select(selector);
-      if (elements.size() != 1) {
-        throw new AssertionError("Expected a single element to match the selector " + selector + ". Found [" + elements.size() + "] instead." + ((elements.size() == 0) ? "" : "\n\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      Element element = elements.get(0);
-      if (!element.val().equals(value.toString())) {
-        throw new AssertionError("Using the selector [" + selector + "] expected [" + value + "] but found [" + element.val() + "]. Actual matched element: \n\n" + element + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Allow for custom assertions on an element.
-     *
-     * @param selector the DOM selector
-     * @param consumer a consumer that will take the element found by the selector
-     * @return this.
-     */
-    public HTMLAsserter assertOnElement(String selector, Consumer<Element> consumer) {
-      Element element = selectExpectOne(selector);
-      consumer.accept(element);
-      return this;
-    }
-
-    /**
-     * Assert that a Select option is not selected.
-     *
-     * @param selector the DOM selector
-     * @return this.
-     */
-    public HTMLAsserter assertOptionIsNotSelected(String selector) {
-      Element element = selectExpectOne(selector);
-
-      if (!element.is("option")) {
-        throw new AssertionError("Expected the element not be an [option] but found [" + element.tagName().toLowerCase(Locale.ROOT) + "].\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      if (element.hasAttr("selected")) {
-        throw new AssertionError("Expected the element not to be selected." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Assert that an option is selected.
-     *
-     * @param selector the DOM selector
-     * @return this.
-     */
-    public HTMLAsserter assertOptionIsSelected(String selector) {
-      Element element = selectExpectOne(selector);
-
-      if (!element.is("option")) {
-        throw new AssertionError("Expected the element not be an [option] but found [" + element.tagName().toLowerCase(Locale.ROOT) + "].\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      if (!element.hasAttr("selected")) {
-        throw new AssertionError("Expected the element to be selected." + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return this;
-    }
-
-    /**
-     * Perform any custom assertions on the parsed HTML document.
-     *
-     * @param consumer the HTML document consumer
-     * @return this.
-     */
-    public HTMLAsserter custom(ThrowingConsumer<Document> consumer) throws Exception {
-      consumer.accept(document);
-      return this;
-    }
-
-    private Element selectExpectOne(String selector) {
-      Elements elements = document.select(selector);
-      if (elements.size() != 1) {
-        throw new AssertionError("Expected a single element to match the selector " + selector + ". Found [" + elements.size() + "] elements instead." + ((elements.size() == 0) ? "" : "\n\n" + elements) + "\n\nActual body:\n" + requestResult.getBodyAsString());
-      }
-
-      return elements.get(0);
-    }
-  }
-
-  public static class JSONAsserter {
-    public String errorMessage;
-
-    public JsonNode node;
-
-    public Integer size;
-
-    public JSONAsserter(JsonNode node) {
-      this.node = node;
-    }
-
-    public JSONAsserter assertLength(int expected) {
-      if (node.size() != expected) {
-        errorMessage = "Expected an array of size [" + expected + "] but found [" + node.size() + "]";
-      }
-
-      return this;
-    }
-
-    public JSONAsserter assertType(JsonNodeType expected) {
-      if (node.getNodeType() != expected) {
-        errorMessage = "Expected node to be of type [" + expected + "] but found [" + node.getNodeType() + "]";
-      }
-
-      return this;
-    }
-  }
-
-  public static class TestURIBuilder extends QueryStringBuilder {
-    private TestURIBuilder() {
-    }
-
-    private TestURIBuilder(String uri) {
-      super(uri);
-    }
-
-    public static TestURIBuilder builder() {
-      return new TestURIBuilder();
-    }
-
-    public static TestURIBuilder builder(String uri) {
-      return new TestURIBuilder(uri);
-    }
-
-    @Override
-    public TestURIBuilder uri(String uri) {
-      return (TestURIBuilder) super.uri(uri);
-    }
-
-    @Override
-    public TestURIBuilder with(String name, Object value) {
-      return (TestURIBuilder) super.with(name, value);
-    }
-
-    @Override
-    public TestURIBuilder with(String name, Consumer<QueryStringBuilder> consumer) {
-      return (TestURIBuilder) super.with(name, consumer);
-    }
-
-    /**
-     * Add a parameter to the request that you will expect to match the expected. This may be useful if a timestamp oro other random data is returned
-     * that is not important to assert on.
-     * <p>
-     * <strong>This is only intended for use during testing.</strong>
-     *
-     * @param name the parameter name
-     * @return this
-     */
-    @Override
-    public TestURIBuilder withActual(String name) {
-      with(name, "___actual___");
-      return this;
-    }
-
-    public TestURIBuilder withConsumer(Consumer<TestURIBuilder> consumer) {
-      if (consumer != null) {
-        consumer.accept(this);
-      }
-
-      return this;
-    }
-
-    /**
-     * Add a parameter as optional which means that it is not required to be on the query string, but if it is, the actual value will be used during
-     * the assertion.
-     * <p>
-     * <strong>This is only intended for use during testing.</strong>
-     *
-     * @param name the parameter name
-     * @return this
-     */
-    public TestURIBuilder withOptional(String name) {
-      with(name, "___optional___");
-      return this;
-    }
-
-    public TestURIBuilder withQueryString(String queryString) {
-      Map<String, List<String>> parsed = QueryStringTools.parseQueryString(queryString);
-      parsed.forEach((name, values) -> values
-          .forEach(value -> super.with(name, value)));
-      return this;
-    }
-
-    @Override
-    public TestURIBuilder withSegment(Object segment) {
-      return (TestURIBuilder) super.withSegment(segment);
-    }
-  }
 }
